// src/app/models/enums/defence-type.ts
export enum DefenceType {
  LIGHT_BEAM_CANNON = 'Light Beam Cannon',
  BEAM_CANNON = 'Beam Cannon',
  HEAVY_BEAM_CANNON = 'Heavy Beam Cannon',
  SAM_SITE = 'SAM Site',
  ORBITAL_MISSILE_LAUNCHER = 'Orbital Missile Launcher',
  HEAVY_ORBITAL_MISSILE_LAUNCHER = 'Heavy Orbital Missile Launcher',
  RAIL_GUN_CANNON = 'Rail Gun Cannon',
  SMALL_BOMB = 'Small Bomb',
  CLUSTER_BOMB = 'Cluster Bomb',
  MEDIUM_BOMB = 'Medium Bomb',
  HEAVY_BOMB = 'Heavy Bomb'
}